admin_portal_backend_service/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pydantic-settings
python-jose[cryptography]